Fix a bug that splices the most popular venues.
When decidig which nearby venues to render, and two venues are too close
to each other, make sure to pick the most popular one to display. This
fixes a bug where the oppossite was happening, choosing the most popular
venues to splice them out (thereby rendering only the least popular
venues).