Add navigation bar links for each city.
Clicking each link will set the view of the map to the center
of the respective city. This is done by the new method
MapApp.centerOn(), which takes the string representing the city
as specified in the Cities object.

MapApp.defaultCenter variable has been replaced by MapApp.defaultArea
variable, which holds the string representing the default city.