Add log comparison in testing framework
Add functionality to retrieve MapApp debug logs from
the testing framework. The JSON logs are then translated
into Java objects of the class LogEntry. To compare two lists
of LogEntry objects generated from the interaction between a
leader user and a listener user, we can take one of the lists
and generate its complement list of actions. The complement
list should be equal to the second list from the leader-listener
interaction.