Fix to eliminate messages received by sender
Improve implementation of pending and received messages
to eliminate the messages that are sometimes received by
a user that only performs send operations.

Modified test to not expect a hardcoded number of log messages.
Instead, it compares between browsers to see if they all have
the same number of messages.

With these changes all tests are passing.